+++ backend/Muhasebe/MuhasebeHesapPlanlari/Repositories/IMuhasebeHesapPlaniRepository.cs
@@ -1,0 +1,8 @@
+using STYS.Muhasebe.MuhasebeHesapPlanlari.Entities;
+using TOD.Platform.Persistence.Rdbms.Repositories;
+
+namespace STYS.Muhasebe.MuhasebeHesapPlanlari.Repositories;
+
+public interface IMuhasebeHesapPlaniRepository : IBaseRdbmsRepository<MuhasebeHesapPlani, int>
+{
+}